Pass a function which returns the dateFormat function to AMD `define`
When requiring an AMD module, you're given the return value of that module definition. In this case, the return value of `dateFormat` was in fact the default format of the current date and time.

We instead should be passing an anonymous function to `define`, which in turn returns the `dateFormat` function.