Add internationalization (i18n) for pt-br (Portuguese Brazilian language)!
There are no separate files for internationalization is the library, so I'm creating the first to be set as the default to be loaded by the browser. It could be changed to library to facilitate the inclusion of new languages and switching between them. I used the example of the library [Numeral-js] (https://github.com/adamwdraper/Numeral-js).

I hope my contribution is appreciated! = D